Revise paper according to reviewer comments
Changed dayahead and intra-day comparison interval to include a much
longer time period over all seasons of the year. Also updated some text
within the paper.

This version includes notation of revisions in the tex file that can be
turned on to highlight changes in the pdf. A further commit will be
prepared that removes these "track changes" type of annotations for
cleaner final tex code.